+++ src/MicroElements/Configuration/Evaluation/IValueEvaluator.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
+
 namespace MicroElements.Configuration.Evaluation
 {
     /// <summary>
@@ -11,7 +13,7 @@ namespace MicroElements.Configuration.Evaluation
         /// <summary>
         /// The name of evaluator.
         /// </summary>
-        string Name { get; }
+        EvaluatorInfo Info { get; }
 
         /// <summary>
         /// Expression evaluation.
@@ -20,5 +22,32 @@ namespace MicroElements.Configuration.Evaluation
         /// <param name="expression">Expression.</param>
         /// <returns>Evaluated result.</returns>
         string Evaluate(string key, string expression);
+    }
+
+    /// <summary>
+    /// Represents evaluator information.
+    /// </summary>
+    public class EvaluatorInfo
+    {
+        /// <summary>
+        /// The name of evaluator.
+        /// </summary>
+        public string Name { get; }
+
+        /// <summary>
+        /// Evaluator order.
+        /// </summary>
+        public int Order { get; }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="EvaluatorInfo"/> class.
+        /// </summary>
+        /// <param name="name">The name of evaluator.</param>
+        /// <param name="order">Evaluator order.</param>
+        public EvaluatorInfo(string name, int order = 100)
+        {
+            Name = name ?? throw new ArgumentNullException(nameof(name));
+            Order = order;
+        }
     }
 }